requirements.txt
click==7.1.2
fastapi==0.63.0
h11==0.12.0
pydantic==1.7.3
starlette==0.13.6
uvicorn==0.13.3
requests==2.25.1
tortoise-orm==0.16.21
pandas==1.2.2
aiohttp==3.7.3
asyncpg==0.22.0
psycopg2-binary==2.8.6
dash==1.19.0
python-dateutil==2.8.1
